Divided main.go into generate.go, postgres.go and main.go.
