+++ pq2g/generate.go
@@ -1,0 +1,91 @@
+package pq2g
+
+import (
+  "strings"
+
+  "github.com/gedex/inflector"
+)
+
+// Infer belongs_to Relation from column's name
+func inferORM(s string) (bool, string) {
+  s = strings.ToLower(s)
+  ss := strings.Split(s, "_")
+
+  newSS := []string{}
+  var containsID bool = false
+  for _, word := range ss {
+    if word == "id" {
+      containsID = true
+      continue
+    }
+
+    newSS = append(newSS, word)
+  }
+
+  if containsID == false || len(newSS) == 0 {
+    return false, ""
+  }
+
+  infColName := strings.Join(newSS, "_")
+  return true, infColName
+}
+
+// Generate json
+func genJSON(columnName, columnDefault string, primaryKeys map[string]bool) (json string) {
+  json = "json:\"" + columnName + "\""
+
+  if primaryKeys[columnName] {
+    p := "gorm:\"primary_key;AUTO_INCREMENT\" "
+    json = p + json
+  }
+
+  if columnDefault != "" && !strings.Contains(columnDefault, "nextval") {
+    d := " sql:\"DEFAULT:" + columnDefault + "\""
+    json += d
+  }
+
+  return
+}
+
+// Singlarlize table name and upper initial character
+func gormTableName(s string) string {
+  s = strings.ToLower(s)
+  s = inflector.Singularize(s)
+  return strings.Title(s)
+}
+
+// Ex: facebook_uid → FacebookUID
+func gormColName(s string) string {
+  s = strings.ToLower(s)
+  ss := strings.Split(s, "_")
+
+  for i, word := range ss {
+    if strings.Contains(word, "id") {
+      word = strings.Replace(word, "id", "ID", -1)
+    }
+
+    if strings.Contains(word, "url") {
+      word = strings.Replace(word, "url", "URL", -1)
+    }
+
+    ss[i] = strings.Title(word)
+  }
+  return strings.Join(ss, "")
+}
+
+func gormDataType(s string) string {
+  switch s {
+  case "integer":
+    return "uint"
+  case "numeric":
+    return "float64"
+  case "character varying", "text":
+    return "string"
+  case "boolean":
+    return "bool"
+  case "timestamp with time zone":
+    return "time.Time"
+  default:
+    return s
+  }
+}
